cosmosdb/fn-requirements.txt
azure-functions
requests
sendgrid